Rename TreeTally to SilviStat
