+++ src/silvistat/shatter.py
